+++ back/swagger_output.json
@@ -1323,6 +1323,40 @@
         }
       }
     },
+    "/admin/bills/{billId}/shopping": {
+      "delete": {
+        "description": "",
+        "parameters": [
+          {
+            "name": "billId",
+            "in": "path",
+            "required": true,
+            "schema": {
+              "type": "string"
+            }
+          }
+        ],
+        "responses": {
+          "200": {
+            "description": "OK"
+          }
+        },
+        "requestBody": {
+          "content": {
+            "application/json": {
+              "schema": {
+                "type": "object",
+                "properties": {
+                  "shoppingIds": {
+                    "example": "any"
+                  }
+                }
+              }
+            }
+          }
+        }
+      }
+    },
     "/catalogs/images": {
       "get": {
         "description": "",
